Componentizes Repository into separate managers
