+++ lib/devlogs/issue_manager.rb
@@ -1,0 +1,3 @@
+class IssueManager
+  ISSUE_FILE_PREFIX = "iss"
+end
